Made a "get profile picture" function
