+++ functions.js
@@ -143,5 +143,51 @@ module.exports = {
 		
 			return resId;
 		}
+	},
+	async getProfilePotho(profileData,TimeOut,Xp,xpPercentage,message){
+		const {createCanvas, loadImage, registerFont} = require("canvas");
+		var whidth = 500;
+        var hight = 800;
+		registerFont("./canvas/Hard Compound.ttf",{family:"Hard_Compound"});
+        const Profile = createCanvas(whidth,hight);
+        const ProfileOptions = Profile.getContext("2d");
+		if(profileData.colorHexCode == undefined){
+			console.log("insde "+ profileData.colorHexCode);
+			profileData.colorHexCode = "#787C75";
+			profileData.save();
+		}
+        ProfileOptions.fillStyle = profileData.colorHexCode;
+        ProfileOptions.fillRect(0,0,whidth,hight);
+        await loadImage(message.author.avatarURL({format:"png"})).then(img =>{
+			ProfileOptions.fillStyle = "#5FDA18";
+			ProfileOptions.fillRect((whidth/2)-135,70,270,270)
+			ProfileOptions.fillStyle = profileData.colorHexCode;
+			ProfileOptions.fillRect((whidth/2)-125,80,250,250)
+            ProfileOptions.drawImage(img,(whidth/2) - 125,80,250,250);
+        })
+        ProfileOptions.font = "bold 50pt Hard_Compound"
+        ProfileOptions.textAlign = "center"
+        ProfileOptions.fillStyle = "#fff"
+        ProfileOptions.fillText(message.author.username,(whidth/2), 400)
+		ProfileOptions.font = "normal 40pt Hard_Compound"
+		ProfileOptions.fillText(`Level: ${profileData.level-1}`,(whidth/2),500);
+		//ProfileOptions.font = "normal 30pt sans"
+		//ProfileOptions.fillText(`Progress:\n${progressBar} ${xpPercentage}%`,(whidth/2),630);
+		var multiplier = 3.5;
+		var fildBar = 100 * multiplier;
+		var Bar = xpPercentage * multiplier;
+		ProfileOptions.fillStyle = "#898C87"
+		ProfileOptions.fillRect(70,550,fildBar,40);
+		ProfileOptions.fillStyle = "#fff"
+		ProfileOptions.fillRect(70,550,Bar,40);
+		ProfileOptions.font = "normal 40pt Hard_Compound";
+		ProfileOptions.fillText(`${xpPercentage}%`,(whidth/2),645);
+		ProfileOptions.font = "normal 20pt Hard_Compound";
+		if(Xp != "") ProfileOptions.fillText(Xp,(whidth/2),700);
+		if(TimeOut != "") ProfileOptions.fillText(TimeOut,(whidth/2) ,730);
+		await loadImage("./canvas/BackrundsFrame.png").then(img =>{
+            ProfileOptions.drawImage(img,0,0,whidth,hight);
+        })
+        return Profile.toBuffer("image/png");
 	}
 }
